mb/google/bluey: Enable boot media detection for Bluey
Move fw_config_override() from mainboard.c to fw_config_bluey.c,
which is compiled only for Bluey.

Populate FW_CONFIG with the detected storage media so the correct
boot storage interface is enumerated during boot.

TEST=Build and boot google/bluey

Reviewed-on: https://review.coreboot.org/c/coreboot/+/95187
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
